Fix Transfer registration page rendering
https://eaflood.atlassian.net/browse/RUBY-933

It was spotted whilst testing some other functionality that the transfer a registration pages were not properly rendering in IE11. A quick check found that the HTML being rendered is not valid because according to http://validator.w3.org/ they are missing some closing tags.

This change covers ensuring the pages render correctly and are valid.