Fix json dependency security issue
https://hakiri.io/technologies/json/issues/894d2260c50062
https://hakiri.io/projects/d675663bf2646a/stacks/ed0cf5abb6046d/builds/e62b03f39ea9b0/warnings/af21802ed6a878

This resolves an issue Hakiri highlighted with the json dependency.

We are not actively using it in our app, it was a depedency of 2 other gems; [Simplecov](https://github.com/colszowka/simplecov) and [sdoc](https://github.com/zzak/sdoc). However, by updating the versions of both to their latest the json depedency was dropped altogether.

Problem solved!

N.B. The simplecov version was originally pinned due to a breaking change between it and CodeClimate's test-reporter.
As we now use SonarCloud this is no longer an issue.

https://github.com/DEFRA/waste-exemptions-engine/pull/421