Add expired registrations in grace window to EPR
From: https://eaflood.atlassian.net/browse/RUBY-982 and https://eaflood.atlassian.net/browse/RUBY-983

This changes the scope for fetching registration for the EPR export to include EXPIRED registrations in the grace window.
It also changes the exported data, so that the expire date is inflated with the grace window value for all expired registrations exported.