Add API endpoint to seed data using JSON seeds
This will be used to improve performances of the Acceptance test suite and will make existing seeds file easely reusable.
The main change in seeding is that this API will generate a new registration identifier and return it as a response, so that the test suite can run multiple times without having to wipe the DB data.