Generate magic link when sending email reminders
From: https://eaflood.atlassian.net/browse/RUBY-986

Generates magic link when sending a first email reminder and when sending a second unless a token already exists.
This also updates the API to renturn the magic link as part of registration information and adds the feature toggle for magic link renewals.