Add one-off rake task to update defunct renewal workflow_state
This task exists to find any renewals with the workflow_state `renewal_received_form` and change them to one of the new workflow_states we've split that into.

We should only need to run this once on production, when we release the updated workflow.

Also updated another rake task that was still using the old workflow_state.