Update templates of first and second renewal email reminders
From: https://eaflood.atlassian.net/browse/RUBY-1019

This update the template of the first and second email reminders to match the magic link template. In the new template, we don't have differences for those email templates. Little refactoring has been done there, as it is quite possible that the email templates will differ again in the near future.