Uodate back button path to new backoffice
From: https://eaflood.atlassian.net/browse/RUBY-1038

We used to redirect the user to the privacy policy page for only renewals, from the old backend.
Since now we are using this page for new registrations that can be triggered from the new backoffice as well, we want the path to redirect always to the new backoffice instead.