Add link to re-send a renewal reminder email
From: 	https://eaflood.atlassian.net/browse/RUBY-1046

This send a renewal reminder email to the user using a button in the registration details page.
Given that the first email reminder will always generate a new renew token, but we don't want to generate a new token if present when we send again a renewal email, and given that the second renewa email has the same template as the first but deals with the creation of the renewal link as needed here, we are using the second renewal email when sending via back office link.