Allow NCCC to send new regs back from Worldpay to payment summary
https://eaflood.atlassian.net/browse/RUBY-1075

Renewals have a feature where if the current state is Worldpay, an NCCC user can click the “Revert from Worldpay” button on the view details page in the back office to kick the renewal back to the payment summary form.

This is useful if there is a Worldpay error, or if the user is otherwise stuck.

It should be possible to do this for new registrations as well.