Allow finance admins to add missed Worldpay payments to in-progress new registrations
https://eaflood.atlassian.net/browse/RUBY-1093

Finance admin users should be able to add a missed Worldpay payment to an in-progress new registration that is stuck in the "worldpay_form" workflow status.

Because our finance forms are built to work with Registrations rather than NewRegistrations, this new controller converts the NewRegistration into a Registration before adding the payment.