Toggle where new registration redirects to
https://eaflood.atlassian.net/browse/RUBY-1118

Now we can control features through toggles in the UI, it has exposed that we do not have a clean separation of functionality when `new_registration` is enabled and disabled.

This change deals with one such issue found. If the new 'new registration' journey feature is enabled, we want all links to the journey to point to the new apps. This means front-office for external users and back-office for internal users.

If it's disabled we want all links to point the old frontend. We have found if the feature is disabled then the back-office will still try to point to the back-office new registration journey. But because the option is disabled it has logic to return a 404 not found.

We instead need to get the back-office to redirect the user to the old frontend if the feature is disabled.