Refactor ad privacy policy helpers to a presenter
There was already a reasonable amount of complexity hidden in the helper, some of which should have been hidden behind a `private`. By adding our additional functionality to the helper we got things working. However, I felt this had reached the point we needed to look at cleaning things up.

So this change refactors the helper to be a presenter, and includes tests to cover the 'new_registration' functionality being both enabled or disabled.