Add new panel and update main view to show it
TBH the logic to determine whether to show the panel or not is too much to have embedded in a view. But I also don't think we should keep adding methods to the `registration` that are used to determine things in views. We really need to refactor stuff to use presenters where they are needed.

So in this case I felt this was the lesser of evils for now but happy to be told 'hell no'!