Add payment confirmation email to Edit
https://eaflood.atlassian.net/browse/RUBY-1160

Now we have the new `receipt_email` field which we are capturing during new registrations and renewals we need to allow users to edit the value. Else once set (or not) it can never be changed which will cause issues for users who need to change it for subsequent Worldpay payments, for example copy card orders.