Remove logic from AD privacy policy
As we are now controlling access to the new registration journey by showing or hiding the main link, there is no need for the logic in the AD presenter anymore.

If you got to the AD privacy policy and you are doing a new registration, we can assume it's because `new_registration` is enabled.