Tell engine we are the back-office
https://eaflood.atlassian.net/browse/RUBY-1141

We have come across an issue where we need to change behaviour in the journey based on whether a registration or renewal is being carried out in either the front-office app or the back-office app.

The simplest way we felt to do this was to add a new flag to the config (see [Add host_is_back_office flag to config](https://github.com/DEFRA/waste-carriers-engine/pull/894)) and set it in the initializer.