Display renewal reminder magic link in reg. detail
https://eaflood.atlassian.net/browse/RUBY-1166

If a registration is eligible for renewal NCCC would like us to display the same renewal link that gets included in the renewal reminder emails and letters in registration details.

However, the link should not be clickable but shown as plain text. This is to avoid a user accidentally clicking on it and starting the renewal.

The reasoning behind it is so NCCC users can manually send the link to someone, for example, a 3rd party agent, who has a different email from the contact email stored in the registration.