Add helper to generate the link
TBH not 100% comfortable with where this is sitting. It's not actually an action link but we depend on the `a_registration?()` in the action links helper. And we know its definately available in the page we need to show the link.

Plus we're duplicating what's already in the RenewalReminderMailer.

However, this is at least a tested woring solution 🤷