Add display_renewal_link_for? helper
Realised that once a registration has a renew token generated, the current functionality would always display it. So we also needed to factor in whether the renewal was actually eligible for renewal.

With all these checks going on it felt time to add another helper rather than add more logic directly into the view.