Merge email_reminders and renew_via_magic_link
https://eaflood.atlassian.net/browse/RUBY-1170

Our intention was to provide automated renewal reminders and build on it with adding magic links and being able to renew through them.

So we added added 2 separate feature toggles to control what was enabled.

In hindsight, the app hasn't actually been built to operate for each of the combinations those 2 feature toggles could be in. Essentially now we consider renewals will only ever be shipped with magic link functionality.

So this change simplifies the logic around renewals by merging these toggles into a new `renewal_reminders` feature toggle.

It also encompass changes to ensure renewal reminder content is correctly enabled and disabled depending on the state of the toggle.