Add missing toggle to feature_toggles.yml
As has been noted previously adding our feature toggles to this file is no longer necessary.

However, having them listed here is a handy way of documenting what features we can toggle. At least until we move the info to somewhere more clear like the README.

So this isn't pertinent to the current change. Its just a bit of housekeeping.