Show err msg when resend reminder to empty contact
https://eaflood.atlassian.net/browse/RUBY-1168

Currently if you click the 'Resend renewal reminder' link on a registration with an empty contact email it will just say 'Renewal sent to'.

With this message you could be mistaken for thinking it sent successfully. Or if you spot it, it wouldn;t be immediately obvious what the issue is.

We now have a design to display an error message if the link is clicked and the contact email is blank. This change implements it.