Update flash message logic
https://eaflood.atlassian.net/browse/RUBY-1168
https://eaflood.atlassian.net/browse/RUBY-1169

When resending the renewal reminder email from the registration details screen it has been flagged its doesn't really work from a UI perspective if the contact email is missing or the AD default.

So we have an update to the design which requires us to show a flash message styled as an error. However, though we have this in WEX we don't appear to have this functionality in WCR.

So this update is to bring WCR in line with WEX and add the ability to show error styled flash messages.