Update crtl. to use new flash concern for message
Update any controllers that currently use `flash[:message]` to use the new concern.

Note. Turns out there is only 1 which is the resend renewal email controller. And actually it should be generating a flash[:error] not a message for the scenario outlined. For now though will just update and comeback later to refactor it.