Don't allow resending/sharing of magic link emails that users can't use
https://eaflood.atlassian.net/browse/RUBY-1175

All the checks for resending emails or otherwise sharing the renewal magic link are based off when NCCC can renew.

However, when NCCC can renew and when a front office user can renew may no longer be the same. In this case, the user could receive a link which would just tell them the registration was expired, and they could no longer use it.

So if the front office can't renew a registration, we shouldn't be sharing links to renew in the front office.