Add exceptions module with first error
Opted to store our exceptions in a central place as per https://stackoverflow.com/a/5369773/6117745.

Advice on Rails 6 now is also to store the stuff you would have previously placed in `lib/` in `app/lib` as per https://github.com/rails/rails/issues/37835

Because we are stil using the classic autoload it seems we needed to manually add `app/lib` to the autoload_path. But in future when the app is ready to switch to the new Zeitwerk loader no other changes should be needed.