Upload EPR exports to EPR folder
https://eaflood.atlassian.net/browse/RUBY-1239

This PR switches to an updated version of the defra-ruby-aws gem to enable us to upload to a subdirectory.

S3 doesn't really have subdirectories so it's actually just appending an "EPR/" prefix to the file on upload.

I also toned down the level of testing as the old tests were written before we moved all the export functionality into the defra-ruby-aws gem. We have much more specific testing in the gem so there is no need to test so heavily here.