Copy cards - remove duplicate business name from house number
https://eaflood.atlassian.net/browse/RUBY-1786

If a registration has a company_name duplicated in the house_number
field, we don't want to display it on the copy cards export.

We also want this to be a case insensitive match, and have updated the
existing filter for company_name in address_line_1 to also  be case
insensitive.