[RUBY-3493] Refactor RenewalReminderServiceBase to return the count of affected registrations
This commit refactors the RenewalReminderServiceBase class to return the count of affected registrations when sending renewal reminder emails. It also adjusts the rake task in lib/tasks/email.rake to log the number of affected registrations using the StdoutLogger class. This change improves the visibility of the number of emails sent and helps with debugging and monitoring.